requirements.txt
openai==1.51.0
python-dotenv==1.0.1
requests==2.32.3
flask_cors==5.0.0
gunicorn